Remove unblocker
It does not work.

It was created for an older era when the user page was most heavily
restricted, and graphql timeline was free. So the visitor would look
up the username-userID relationship on the instance's behalf, and
submit that for the instance to check, and then that profile would be
unblocked forever because the user page is not needed after that
point.

Now, the user page is free, and graphql timeline can be impossible.
(Still haven't worked that out yet.) So the unblocker would only be
fetching information that the instance could already get. Even if the
instance was somehow blocked from the user page, the unblocker would
not help, since it only fetches the username-userID relationship for
use with graphql timeline, and graphql timeline is currently blocked
on the instance too.

Keeping this in Bibliogram is misleading to visitors and the backing
code is now useless.

The correct way to view profiles is to run your own Bibliogram.