themes: Fix incorrect blur and opacity interaction
Chrome sometimes seems to have problems when using a blur
backdrop-filter in combination with opacity. On Linux, this often
results in the blur being completely ignored. This also seems to apply
to other systems, though. See the following issue for more details:
https://bugs.chromium.org/p/chromium/issues/detail?id=1129838

Making the background opaque using rgba seems to fix the problem.