Set 'str_getcsv' escape param to empty string to avoid PHP 8.4 deprecation message.
The escape param is in the process of being eliminated, with PHP 8.4 deprecating passing anything but an empty string as its value.
For some reason they're leaving the default value (a backslash) as-is, meaning the default will cause a deprecation message.
This commit avoids that by setting the escape param to an empty string (see references below).

* https://wiki.php.net/rfc/deprecations_php_8_4#deprecate_proprietary_csv_escaping_mechanism
* https://www.php.net/manual/en/function.str-getcsv.php
* https://old.reddit.com/r/PHP/comments/1eyum8c/new_deprecation_notices_in_php84_with_csv/
* https://nyamsprod.com/blog/csv-and-php8-4/