Store sqlite at /fmd/db/fmd.sqlite
In Docker, it is easiest to mount directories, (like "/fmd/db/").
Because to mount a file, the file needs to exist.
Otherwise Docker may interpret "/fmd/db/fmd.sqlite" as a directory!

Thus for users/admins it is easiest to mount a directoy and let FMD Server
create the sqlite db.