Fix: handle IsLocked inside RequestAccess instead of in HTTP handler
This also fixes https://gitlab.com/Nulide/findmydeviceserver/-/issues/67.
This nil crash + HTTP 502 was due to setCommandToUser being called for a
non-existent user. Now we only check IsLocked and only call setCommandToUser
if the user exists.

As a side-effect, this means that (unlike before) a bot can now hammer log-in
attempts for a non-existent ID. Nevertheless, this is fine because:

- There is no security/privacy risk (the ID doesn't exist).
- Rate limiting for normal DoS mitigation should be done on a per-IP level anyway.