Tweak location metadata view
Motivation: I want the <p> elements in the privacy policy to create line breaks.
This is prevented by `display: inline`.

However, removing `display: inline` causes the location metadata to also break,
into multiple lines. To fix that, use <span> instead of <p>.

While we are at it, move the date and time on the same line as the provider.