Revert to using CWD as default web/db dirs
As pointed out by @rlees85, /var and /usr are only writable by root, and thus not good defaults.
See https://gitlab.com/Nulide/findmydeviceserver/-/merge_requests/105#note_2482982422

For the config.yml Viper can handle multiple defaults, thus we can keep /etc in there.
It will prefer /etc if the files exists, but silently fall back to CWD.

Also document this in the README.

Docker
------

Keep using the Unix paths in the Dockerfile, for best practice.
(Even if this comes as a breaking change and forces users to update their mounts.)

Changing this now in one swoop allows us to e.g. migrate the Dockerfile
to a native Debian package in the future.