feat: Use go:embed for frontend
This patch switches from the existing path parameter to an embedded web
directory. This change makes it possible to deliver fmd-server as a
single binary as well as offering the ability to use tools like ko for
container builds, without adjusting the app further to adhere ko
conventions, simplifying container builds.