Add last_seen_time to fmd_user table
Feature-wise, this is a prerequisite for
https://gitlab.com/fmd-foss/fmd-server/-/issues/82

Operations-wise, this is a useful data point for server admins
who want to clean up and delete old accounts.

Note: Server admins still need to be careful with deletion,
since FMD Android may not contact FMD Server in a long time.
This is expected, if the Android-side connectivity check is disabled and the user never needs FMD.