Add info text about missing auto-refresh
A lot of people are confused by the lack of auto-refresh.
They are used to things _just working_. Where is my location?!?

Until we implement auto-refesh, let's add a note to make this more transparent to users.
It's easy to do once you know it, but it may not be obvious a priori.