Error message when cancelling a label creation
When "Create label" menu is clicked, a popup asks for the label name. If the Cancel button is immediately entered, an error message is displayed: TypeError: can't access property "trim", caption is null.

Indeed, the JavaScript prompt() message returns null when Cancel is pressed. The test "caption !== undefined" is thus true, and the trim() is tried on null, which fails.

The fix is to only check for "caption" or "caption != null".