When a filter rule is not inversed display it in green
When a filter is edited, the Match tab contains one line per rule.

Rules which are not inverted should appear in green color. However, every rule line appears in red.

This is due to $rrow['inverse'] not being unset before calling _get_rule_name().

Indeed, function _get_rule_name() expects $rrow['inverse'] to be unset to mean "false", in order to have a behavior similar to the HTML checkbox which is unset when not checked on the web page.