Remove the old .app_is_ready file earlier
The $DST_DIR/.app_is_ready indicates the app is ready. The previous file should be removed sooner, instead of being removed in the middle of startup.sh.

It is used by updater.sh, which waits sufficiently long to start (30 seconds), and then waits again watching the above file and config.

If the user creates docker-compose.override.yml to set a post_start command, he/she can also wait for this file. Removing it sooner allows for a shorter waiting time.