feat: Stick with TLS intention instead of failing to insecure port
This patch changes the behaviour of the TLS port setup. Instead of
falling back to the insecure port if either the certificate or key file
are not found, this patch adjusts the behaviour to make the server fail
explicitly.

To stay mostly compatible with the existing behaviour, it checks if
either of the config parameters for TLS are set.

As a result this change should only break instances that have a TLS
certificate or key configured that doesn't exist.