When adding tags to an article process linefeeds like commas
When a user adds tags to an article, a textarea is used to retrieve the values. The form indicates that tags have to be separated by commas.

However, it is also convenient to separate them by linefeeds. A user can also enter a linefeed thinking it is a separator.

Keeping tags with new lines in the middle is not useful. So, split the string on commas and linefeeds.