Search the key of an invalid key pair as Full Text Search
A known issue was that the Search Query _title_ was searched using LIKE instead of Full Text Search. It was the same for all keys of an invalid key pair. Solving this suppresses a lot of redundant code.

The Search Queries _start:false_, _pub:false_ and _unread:false_ are now checked.

If the check fails, the SQL error message is bold/strong to be more visible.

The Search Query _@invalid-date_ is now also searched as Full Text Search. Unfortunately, there is no error message, but document it in the source code.