In title/author/note searches support spaces to match exact words
A search filter can for example contain spaces such as _title:" be "_.

In this case, the user as explicitly added spaces (instead of using _title:be_), so she clearly indicates her intent to search the word _be_, and not _cyBErspace_.

This patch supports these cases, for keywords title, author and notes.

Also, do a _mb_strtolower()_ globally, because it was in every keyword processing.

Also process the double quoting magic with a consistent regexp in the three needed cases. This change was needed because the previous regexp did not support leading spaces in the value part of a keyword pair.