Automatically close the "Opening debugger" message
In the Preferences page, in the Feeds tab, in "Feeds with errors" page, one can select a feed, and click on the button "Debug selected feeds".

This opens a new web browser tab, and also adds a yellow message "Opening debugger for selected feeds...". However, when we come back to the Preferences page, we have to manually close this yellow message, which becomes useless.

This patch automatically closes the message after 5 seconds, by setting the second parameter of Notify.progress() to false.

This patch also opens the confirm() only when there are several selected feeds. There is not need for a confirmation for one feed only.